(bzr r3320.1.3)

+++ launcher/LauncherIcon.h
@@ -124,7 +124,7 @@ public:
     return false;
   }
 
-  virtual unsigned long long SwitcherPriority()
+  virtual uint64_t SwitcherPriority()
   {
     return 0;
   }
