+++ unity-shared/FileManager.h
@@ -24,6 +24,7 @@
 #include <memory>
 #include <vector>
 #include <string>
+#include <vector>
 #include <sigc++/sigc++.h>
 
 namespace unity
@@ -37,14 +38,14 @@ public:
   FileManager() = default;
   virtual ~FileManager() {}
 
-  virtual void Open(std::string const& uri, unsigned long long timestamp = 0) = 0;
-  virtual void OpenActiveChild(std::string const& uri, unsigned long long timestamp = 0) = 0;
-  virtual void OpenTrash(unsigned long long timestamp) = 0;
+  virtual void Open(std::string const& uri, uint64_t timestamp = 0) = 0;
+  virtual void OpenActiveChild(std::string const& uri, uint64_t timestamp = 0) = 0;
+  virtual void OpenTrash(uint64_t timestamp) = 0;
   virtual std::vector<std::string> OpenedLocations() const = 0;
   virtual bool IsPrefixOpened(std::string const& uri) const = 0;
   virtual bool IsTrashOpened() const = 0;
   virtual bool IsDeviceOpened() const = 0;
-  virtual void EmptyTrash(unsigned long long timestamp = 0) = 0;
+  virtual void EmptyTrash(uint64_t timestamp = 0) = 0;
 
   sigc::signal<void> locations_changed;
